actualizacion codigo para nueva version de php.

+++ controSubida.php
@@ -1,8 +1,7 @@
-<? session_save_path("sesiones");
+<?php session_save_path("sesiones");
 session_start();
 if($_SESSION['nrpresta'] == null)
-	header ("Location: http://www.ospim.com.ar/prestadores/caducaSes.php");
-	
+	header ("Location: loginPresta.php?err=2");
 include ("conexion.php");
 ?>
 
@@ -24,20 +23,19 @@ body {
 -->
 </style>
 </head>
-<?
+<?php
 	$error=0;
-	$datos = array_values($HTTP_POST_VARS);
+	$datos = array_values($_POST);
 	$pres=$datos[0];
 	$mes=$datos[1];
 	$anio=$datos[2];
 	$tit=$datos[3];
 	$fam=$datos[4];
-	$arc=$archivo_name;
+	$arc=$_FILES['archivo']['name'];
+	$archivo=$_FILES['archivo']['tmp_name'];
 	$preArc=(int)substr($arc,0,3);
 	$mesArc=(int)substr($arc,3,2);
 	$anioArc=(int)substr($arc,5,4);
-	
-	
 ?>
 <body>
 <div align="center"><span class="Estilo3">CONTROL DATOS SUBIDA  </span> </div>
@@ -47,7 +45,7 @@ body {
     <tr>
       <td width="133"><em><strong>Mes</strong></em></td>
       <td width="234"><div align="center">
-        <?
+        <?php
 	  	if ($mes == 0) {
 			print("ERROR");
 			$error=1;
@@ -62,7 +60,7 @@ body {
       <td><em><strong>A&ntilde;o</strong></em></td>
       <td> 
 	    <div align="center">
-	      <?
+	      <?php
 	  	if ($anio == 0) {
 			print("ERROR");
 			$error=1;
@@ -77,7 +75,7 @@ body {
       <td><em><strong>Titulares</strong></em></td>
       <td>
 	    <div align="center">
-	      <?
+	      <?php
 	  	if (!is_numeric($tit)){
 			print("ERROR");
 			$error=1;
@@ -91,7 +89,7 @@ body {
       <td><em><strong>Familiares</strong></em></td>
       <td>
 	    <div align="center">
-	      <?
+	      <?php
 	  	if (!is_numeric($fam)){
 			print("ERROR");
 			$error=1;
@@ -105,7 +103,7 @@ body {
       <td><em><strong>Archivo</strong></em></td>
       <td>
 	  	<div align="center">
-	  	  <?
+	  	  <?php
 		  	//para mostrar el prestador...
 		  	$arcOK=0;
 			if ($preArc!=$pres) {
@@ -125,7 +123,7 @@ body {
 			}
 			if ($arcOK == 0) {
 				$sql = "select * from usuarios where codigo=$pres";
-				$result = mysql_db_query("uv0471_prestador",$sql,$db);
+				$result = mysql_query($sql,$db);
 				$row = mysql_fetch_array($result);
 				print($row['nombre']);
 			}
@@ -135,51 +133,45 @@ body {
     </tr>
   </table>
   <p>
-  <?
+  <?php
   	if ($error==1) {
 		print ("SE PRODUJO UN ERROR - NO SE HA SUBIDO EL ARCHIVO<br />");
 	} else {
-		$destino=$pres."C23".$pres."/".$archivo_name;
-		copy($archivo,$destino);
-		
-		
-		$fecSub=date("Y-m-d");
-		$horSub=date("H:i:s");
-		$totBen=$tit+$fam;
-		
-		//VERFICO SI EXISTE Y ELIMINO
-		$sql2 = "select * from subida where codigo=$pres and mespad=$mes and anopad=$anio";
-		$result2 = mysql_db_query("uv0471_prestador",$sql2,$db);
-		if (mysql_num_rows($result2) == 1) {
-			$sql3 = "delete from subida where codigo=$pres and mespad=$mes and anopad=$anio";
-			$result3 = mysql_db_query("uv0471_prestador",$sql3,$db);
-			if ($result3 == 1) {
-				print("EL REGISTRO DE SUBIDA YA EXISTIA. SE REEMPLAZARA CON LOS DATOS ACTUALES.<br />");
-			} else {
-				print("ERROR AL QUERER ELIMINAR REGISTRO EXISTENTE<br />");
+		$destino=$pres."C23".$pres."/".$arc;
+		if (copy($archivo,$destino)) {
+			$fecSub=date("Y-m-d");
+			$horSub=date("H:i:s");
+			$totBen=$tit+$fam;
+			
+			//VERFICO SI EXISTE Y ELIMINO
+			$sql2 = "select * from subida where codigo=$pres and mespad=$mes and anopad=$anio";
+			$result2 = mysql_query($sql2,$db);
+			if (mysql_num_rows($result2) == 1) {
+				$sql3 = "delete from subida where codigo=$pres and mespad=$mes and anopad=$anio";
+				$result3 = mysql_query($sql3,$db);
+				if ($result3 == 1) {
+					print("EL REGISTRO DE SUBIDA YA EXISTIA. SE REEMPLAZARA CON LOS DATOS ACTUALES.<br />");
+				} else {
+					print("ERROR AL QUERER ELIMINAR REGISTRO EXISTENTE<br />");
+				}
 			}
-		}
-		
-		//SUBO EL NUEVO REGISTRO....
-		$sql1 = "INSERT INTO subida VALUES ('".trim($pres)."','".trim($mes)."','".trim($anio)."','".trim($fecSub)."','".trim($horSub)."','".trim($tit)."','".trim($fam)."','".trim($totBen)."','N')";
-		$result1 = mysql_db_query("uv0471_prestador",$sql1,$db);
-		
-		
-		if ($result1 == 1) {
-			print ("SE SUBIO EL ARCHIVO CORRECTAMENTE<br />");
+			
+			//SUBO EL NUEVO REGISTRO....
+			$sql1 = "INSERT INTO subida VALUES ('".trim($pres)."','".trim($mes)."','".trim($anio)."','".trim($fecSub)."','".trim($horSub)."','".trim($tit)."','".trim($fam)."','".trim($totBen)."','N')";
+			$result1 = mysql_query($sql1,$db);
+			if ($result1 == 1) {
+				print ("SE SUBIO EL ARCHIVO CORRECTAMENTE<br />");
+			} else {
+				print ("ERROR AL QUERER ESCRIBIR LA BASE<br />");
+			}
 		} else {
-			print ("ERROR AL QUERER ESCRIBIR LA BASE<br />");
+			print ("SE PRODUJO UN ERROR AL COPIAR EL ARCHIVO - NO SE HA SUBIDO EL ARCHIVO<br />");
 		}
 
-	}
-	
-  ?>
-  </p>
-  <p>
-    <?
- 	 print ("<font face=Verdana size=3><b><font color=#CF8B34><a href=subidaPadron.php>"."VOLVER"."</font></b></font>");
-    ?>
-  </p>
+	}?>
+</p>
+<p>	<input type="button" name="salir" value="VOLVER" onclick="location.href='subidaPadron.php'"/></p>
+
 </div>
 </body>
 </html>
